+++ process_windows.go
@@ -1,0 +1,7 @@
+// +build windows
+
+package wormhole
+
+func shellArgs(program string) []string {
+	return []string{"cmd.exe", "/C", program}
+}